Move fs-tag from mount to mkfs
The fs-tag makes more sense in the mkfs step, because that's
when the filesystem is created.
There is no need to create a new tag in a mount step, because a
filesystem can only be mounted once, so we can find the mountpoint
from the fs-tag.

Also this allows us to mount filesystems that are not on a partition,
but on e.g. a device directly.